Link README.md
Hi it's almost noon, EST, April 8th & you know what this finally irks me
enough to forge this commit. For a while raw-text was kind of chill but
let's see. I reserve the right to back this symlink out as ex-post as I
put it in.

I'm forging the date because I don't want to blip the project.